+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New `MERRA2Dataset` method `to_dataframe()` to convert into a pandas `DataFrame`
 
 ## [0.2.1] - 2019-01-03
 ### Changed